server: S13 parity fixes — messaging, lifecycle, water-timer (batch 1)
Seven CONFIRMED divergences from the S13 legacy-parity fan-out (the four
parallel auditors over lifecycle / physics / world / snapshot), each verified
against the legacy source before fixing.  The higher-risk matrix/rotated-brush
Q-18 finding is handled separately.

Messaging (clients/messages.cpp, abi/engine_table.cpp):
- pfnWriteAngle: legacy (sv_game.c:2794) is a PLAIN truncate
  `((int)(v*256/360)) & 255` written as a char — NOT a wrapped bit-angle.  The
  port used write_bit_angle (normalizes to [0,360)), so negative angles diverged
  (-1.0 → byte 0 legacy vs 255 port).  Restored the plain-truncate form.
- MessageEnd: added the svc_finale/svc_cutscene empty-body null-string fixup
  (sv_game.c:2719-2724) — an empty finale/cutscene now gets its trailing 0 byte
  so the client reads an empty string, not the next stream byte.  Vendored
  k_svc_finale=31 / k_svc_cutscene=34.
- pfnWriteEntity: restored the out-of-range Host_Error (sv_game.c:2837) — an
  entnumber <0 or >=numEntities now routes through the Q-5 host-error surface.

Lifecycle (server.cpp, lifecycle/spawn.cpp):
- exec_load_level: added the SV_SetStringArrayMode(false) reset (sv_init.c:1105)
  before spawn — deactivate leaves the string pool in DYNAMIC mode, so without
  this the 2nd+ map's entity string_t's landed in (and were later clobbered
  within) the runtime-churn dynamic half.  Real corruption on map change.
- activate_server: memset(oldmovevars, 0) instead of memcpy (sv_init.c:644) —
  zeroes the delta baseline so the next SV_UpdateMovevars emits the FULL
  movevars delta after activate (a memcpy suppressed it; latent until the S9
  send lands, but the persistent state was already inverted).
- spawn_server: default an empty hostname to pfnGetGameDescription
  (sv_init.c:970) — wire-observable in A2S/info query replies.
- setup_clients: marked the missing ClearBits(FCVAR_CHANGED) after the maxplayers
  latch (sv_init.c:833) as a tracked stub — needs a cmd_cvar clear-changed
  accessor; low impact (nothing polls that bit at the milestone).

Physics (physics.cpp water_move, Q-18):
- The drown/lava/slime timers compared/stored against a pre-rounded `float`
  clock; legacy uses the DOUBLE sv.time directly.  Changed `now` to double
  (float field promoted for the compare; `now + Nf` in double, rounded once on
  the float store) so air_finished/pain_finished/dmgtime are ULP-exact.

88/88, build green.  Remaining S13 items: the utilities/matrix.cpp double-
intermediate Q-18 finding (rotated SOLID_BSP/PORTAL — physics push_rotate +
world traces/contents) and the PLAUSIBLE set (qport match, datagram headroom,
MSG_SPEC routing, lightstyle length, sv_cheats unlock, per-spawn Delta_Init).

+++ xash3dpp/src/server/physics/physics.cpp
@@ -1305,7 +1305,12 @@ void water_move( ServerRuntime &rt, abi::edict_t *ent ) noexcept
     const int   waterlevel = v.waterlevel();
     const int   watertype  = v.watertype();
     const int   flags      = v.flags();
-    const float now        = static_cast<float>( rt.level.time );
+    // Q-18: the timers compare/store against the DOUBLE server clock (legacy
+    // uses `sv.time` directly — float field promoted to double for the compare,
+    // `sv.time + Nf` computed in double then rounded once on the float store).
+    // A pre-rounded float clock here would do the arithmetic in single precision
+    // and drift by a ULP, flipping a drown/lava boundary a frame early/late.
+    const double now       = rt.level.time;
 
     if ( ( flags & ( abi::k_fl_immune_water | abi::k_fl_godmode ) ) == 0 )
     {
@@ -1319,12 +1324,12 @@ void water_move( ServerRuntime &rt, abi::edict_t *ent ) noexcept
                 if ( dmg < 15.0f )
                     dmg = 10.0f; // quake1 original code
                 v.set_dmg( dmg );
-                v.set_pain_finished( now + 1.0f );
+                v.set_pain_finished( static_cast<float>( now + 1.0f ) );
             }
         }
         else
         {
-            v.set_air_finished( now + 12.0f );
+            v.set_air_finished( static_cast<float>( now + 12.0f ) );
             v.set_dmg( 2.0f );
         }
     }
@@ -1336,7 +1341,7 @@ void water_move( ServerRuntime &rt, abi::edict_t *ent ) noexcept
             // XASH3DPP-STUB(chunk6-S9): EntityWaterExit splash sound.
             v.set_flags( flags & ~abi::k_fl_inwater );
         }
-        v.set_air_finished( now + 12.0f );
+        v.set_air_finished( static_cast<float>( now + 12.0f ) );
         return;
     }
 
@@ -1344,13 +1349,14 @@ void water_move( ServerRuntime &rt, abi::edict_t *ent ) noexcept
     {
         if ( ( flags & ( abi::k_fl_immune_lava | abi::k_fl_godmode ) ) == 0 &&
              v.dmgtime() < now )
-            v.set_dmgtime( v.radsuit_finished() < now ? now + 0.2f : now + 1.0f );
+            v.set_dmgtime( static_cast<float>(
+                v.radsuit_finished() < now ? now + 0.2f : now + 1.0f ) );
     }
     else if ( watertype == ml::k_contents_slime )
     {
         if ( ( flags & ( abi::k_fl_immune_slime | abi::k_fl_godmode ) ) == 0 &&
              v.dmgtime() < now && v.radsuit_finished() < now )
-            v.set_dmgtime( now + 1.0f );
+            v.set_dmgtime( static_cast<float>( now + 1.0f ) );
     }
 
     if ( !fbit( flags, abi::k_fl_inwater ) )
